requirements.txt
mysql-connector-python
google-generativeai